Lots of refactoring and structuring... but also... added a "trigger_pulled" property to the PlayerControl component.
