+++ src/systems/camera.rs
@@ -1,0 +1,18 @@
+use bevy::prelude::{Query, Transform, With, Without};
+use bevy::math::Vec3;
+use crate::components::{CameraFollow, GameCam};
+
+pub fn camera_follow(
+    to_follow: Query<&Transform, (With<CameraFollow>, Without<GameCam>)>,
+    mut camera: Query<&mut Transform, (With<GameCam>, Without<CameraFollow>)>,
+) {
+    let Ok(player_position) = to_follow.get_single() else { return; };
+    let Ok(mut camera_transform) = camera.get_single_mut() else { return; };
+    let target = Vec3 {
+        x: player_position.translation.x,
+        y: player_position.translation.y,
+        z: camera_transform.translation.z,
+    };
+
+    camera_transform.translation = camera_transform.translation.lerp(target, 0.5);
+}
