+++ src/systems/mod.rs
@@ -1,58 +1,12 @@
 pub(crate) mod input;
 pub(crate) mod startup;
+pub(crate) mod camera;
+pub(crate) mod movement;
 
-use crate::components::{CameraFollow, DirectionControl, GameCam, QuadCoord, QuadStore};
-use bevy::math::Vec3;
-use bevy::prelude::{Entity, Query, ResMut, Transform, With, Without};
+use crate::components::{QuadCoord, QuadStore};
+use bevy::prelude::{Entity, Query, ResMut};
 use bevy::utils::HashSet;
-use bevy_xpbd_2d::components::{
-    ExternalForce, Position};
-use bevy_xpbd_2d::prelude::LinearVelocity;
-use crate::boids::{Boid, BoidDirection};
-use crate::components::player::Player;
-
-pub fn camera_follow(
-    to_follow: Query<&Transform, (With<CameraFollow>, Without<GameCam>)>,
-    mut camera: Query<&mut Transform, (With<GameCam>, Without<CameraFollow>)>,
-) {
-    let Ok(player_position) = to_follow.get_single() else { return; };
-    let Ok(mut camera_transform) = camera.get_single_mut() else { return; };
-    let target = Vec3 {
-        x: player_position.translation.x,
-        y: player_position.translation.y,
-        z: camera_transform.translation.z,
-    };
-
-    camera_transform.translation = camera_transform.translation.lerp(target, 0.5);
-}
-
-#[allow(dead_code)]
-fn external_force_player_control(
-    mut query: Query<(&mut ExternalForce, &DirectionControl), With<Player>>
-) {
-    if let Ok((mut external_force, direction_control)) = query.get_single_mut() {
-        external_force.apply_force(direction_control.direction * direction_control.force_scale);
-    }
-}
-
-pub fn linear_velocity_control_boid(
-    mut query: Query<(&mut LinearVelocity, &BoidDirection), (With<Boid>, Without<Player>)>
-) {
-    let mut iter = query.iter_mut();
-    while let Some((mut linear_velocity, direction_control)) = iter.next() {
-        linear_velocity.x = direction_control.direction.x * direction_control.force_scale;
-        linear_velocity.y = direction_control.direction.y * direction_control.force_scale;
-    }
-}
-
-pub fn linear_velocity_control_player(
-    mut query: Query<(&mut LinearVelocity, &DirectionControl), (Without<Boid>, With<Player>)>
-) {
-    if let Ok((mut linear_velocity, direction_control)) = query.get_single_mut() {
-        linear_velocity.x = direction_control.direction.x * direction_control.force_scale;
-        linear_velocity.y = direction_control.direction.y * direction_control.force_scale;
-    }
-}
+use bevy_xpbd_2d::components::Position;
 
 pub fn naive_quad_system(
     mut query: Query<(Entity, &Position, &mut QuadCoord)>,
