+++ src/systems/movement.rs
@@ -1,0 +1,33 @@
+use bevy::prelude::{Query, With, Without};
+use bevy_xpbd_2d::components::{ExternalForce, LinearVelocity};
+use crate::boids::{Boid, BoidDirection};
+use crate::components::PlayerControl;
+use crate::components::player::Player;
+
+#[allow(dead_code)]
+fn external_force_player_control(
+    mut query: Query<(&mut ExternalForce, &PlayerControl), With<Player>>
+) {
+    if let Ok((mut external_force, direction_control)) = query.get_single_mut() {
+        external_force.apply_force(direction_control.direction * direction_control.force_scale);
+    }
+}
+
+pub fn linear_velocity_control_boid(
+    mut query: Query<(&mut LinearVelocity, &BoidDirection), (With<Boid>, Without<Player>)>
+) {
+    let mut iter = query.iter_mut();
+    while let Some((mut linear_velocity, direction_control)) = iter.next() {
+        linear_velocity.x = direction_control.direction.x * direction_control.force_scale;
+        linear_velocity.y = direction_control.direction.y * direction_control.force_scale;
+    }
+}
+
+pub fn linear_velocity_control_player(
+    mut query: Query<(&mut LinearVelocity, &PlayerControl), (Without<Boid>, With<Player>)>
+) {
+    if let Ok((mut linear_velocity, direction_control)) = query.get_single_mut() {
+        linear_velocity.x = direction_control.direction.x * direction_control.force_scale;
+        linear_velocity.y = direction_control.direction.y * direction_control.force_scale;
+    }
+}
